extracted out logic for building commits
