+++ collectors/scm/gitlab/src/main/java/com/capitalone/dashboard/collecteur/DefaultGitlabGitClient.java
@@ -13,11 +13,6 @@ import javax.net.ssl.X509TrustManager;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.joda.time.DateTime;
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -46,13 +41,17 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
     private final RestOperations restOperations;
     private final GitlabUrlUtility gitlabUrlUtility;
     private final GitlabSettings gitlabSettings;
+    private final GitlabResponseMapper responseMapper;
     
     @Autowired
-    public DefaultGitlabGitClient(GitlabUrlUtility gitlabUrlUtility, GitlabSettings gitlabSettings,
-                                       Supplier<RestOperations> restOperationsSupplier) {
+    public DefaultGitlabGitClient(GitlabUrlUtility gitlabUrlUtility, 
+    								   GitlabSettings gitlabSettings,
+                                       Supplier<RestOperations> restOperationsSupplier, 
+                                       GitlabResponseMapper responseMapper) {
         this.gitlabUrlUtility = gitlabUrlUtility;
         this.gitlabSettings = gitlabSettings;
         this.restOperations = restOperationsSupplier.get();
+        this.responseMapper = responseMapper;
     }
 
     @Override
@@ -68,12 +67,9 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
 			int nextPage = 1;
 			while(!lastPage) {
 				ResponseEntity<String> response = makeRestCall(apiUrl, apiToken);
-				JSONArray jsonArray = paresAsArray(response);
-				for (Object item : jsonArray) {
-					JSONObject jsonObject = (JSONObject) item;
-					commits.add(buildCommit(jsonObject, repo.getRepoUrl(), repo.getBranch()));
-				}
-				if(isLastPage(jsonArray.size())) 
+				List<Commit> pageOfCommits = responseMapper.mapResponse(response.getBody(), repo.getRepoUrl(), repo.getBranch());
+				commits.addAll(pageOfCommits);
+				if(isLastPage(pageOfCommits.size())) 
 					lastPage = true;
 				else {
 					apiUrl = gitlabUrlUtility.updatePage(apiUrl, nextPage);
@@ -92,25 +88,6 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
 		if(resultSize < RESULTS_PER_PAGE) 
 			return true;
 		return false;
-	}
-
-	private Commit buildCommit(JSONObject jsonObject, String repoUrl, String repoBranch) {
-		String author = str(jsonObject, "author_name");
-		String message = str(jsonObject, "message");
-		String id = str(jsonObject, "id");
-		long timestamp = new DateTime(str(jsonObject, "created_at")).getMillis();
-
-		Commit commit = new Commit();
-		commit.setTimestamp(System.currentTimeMillis());
-		commit.setScmUrl(repoUrl);
-		commit.setScmBranch(repoBranch);
-		commit.setScmRevisionNumber(id);
-		commit.setScmAuthor(author);
-		commit.setScmCommitLog(message);
-		commit.setScmCommitTimestamp(timestamp);
-		commit.setNumberOfChanges(1);
-		// TODO: figure out commit type and parents
-		return commit;
 	}
 
 	private ResponseEntity<String> makeRestCall(URI url, String apiToken) {
@@ -143,19 +120,5 @@ public class DefaultGitlabGitClient implements  GitlabGitClient {
 		} catch (final Exception ex) {
 			LOG.error(ex.getMessage());
 		}
-	}
-
-	private JSONArray paresAsArray(ResponseEntity<String> response) {
-		try {
-			return (JSONArray) new JSONParser().parse(response.getBody());
-		} catch (ParseException pe) {
-			LOG.error(pe.getMessage());
-		}
-		return new JSONArray();
-	}
-
-	private String str(JSONObject json, String key) {
-		Object value = json.get(key);
-		return value == null ? null : value.toString();
 	}
 }
